post associated to handle delete tag and category

+++ app/Http/Controllers/TagsController.php
@@ -95,6 +95,13 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+
+        if ($tag->posts->count() > 0) {
+            session()->flash('error', 'Tag cannot be deleted because it associated some posts.');
+
+            return redirect()->back();
+        }
+
         $tag->delete();
 
         session()->flash('success', 'Tag deleted successfully.');
